rename folders as well
