+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnInfo/QueryColumnInfo.styled.tsx
