+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnInfo/QueryColumnInfo.tsx
