+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnInfo/QueryColumnInfo.unit.spec.tsx
