+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnInfo/index.ts
