+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnInfoIcon/QueryColumnInfoIcon.tsx
