+++ frontend/src/metabase/common/components/QueryColumnPicker/QueryColumnPicker.tsx
@@ -9,7 +9,7 @@ import {
 import {
   HoverParent,
   QueryColumnInfoIcon,
-} from "metabase/common/components/MetadataInfo/ColumnInfoIcon";
+} from "metabase/common/components/MetadataInfo/QueryColumnInfoIcon";
 import { useLocale } from "metabase/common/hooks";
 import { getColumnGroupIcon } from "metabase/common/utils/column-groups";
 import { useTranslateContent } from "metabase/content-translation/hooks";
