+++ frontend/src/metabase/query_builder/components/view/sidebars/QuestionInfoSidebar/components/SidesheetCardWithFields.tsx
@@ -1,6 +1,6 @@
 import { c, msgid } from "ttag";
 
-import { QueryColumnInfoIcon } from "metabase/common/components/MetadataInfo/ColumnInfoIcon";
+import { QueryColumnInfoIcon } from "metabase/common/components/MetadataInfo/QueryColumnInfoIcon";
 import { SidesheetCard } from "metabase/common/components/Sidesheet";
 import { Box, Group, Stack } from "metabase/ui";
 import * as Lib from "metabase-lib";
