+++ frontend/src/metabase/querying/components/expressions/Editor/Listbox/Listbox.tsx
@@ -14,7 +14,7 @@ import {
 } from "react";
 import { c } from "ttag";
 
-import { QueryColumnInfoIcon } from "metabase/common/components/MetadataInfo/ColumnInfoIcon";
+import { QueryColumnInfoIcon } from "metabase/common/components/MetadataInfo/QueryColumnInfoIcon";
 import type { ExpressionSuggestion } from "metabase/querying/expressions";
 import { Box, DelayGroup, Icon } from "metabase/ui";
 import type * as Lib from "metabase-lib";
